[Bronze V] Title: 검증수, Time: 120 ms, Memory: 14144 KB -BaekjoonHub

+++ 백준/Bronze/2475. 검증수/검증수.java
@@ -3,10 +3,10 @@ import java.io.*;
 class Main {
     public static void main(String[] args) throws IOException {
 
-        long sum = 0;
+        int sum = 0;
         for (int i = 0; i < 5; i++) {
             int n = read();
-            sum += (long) n * n;
+            sum += n * n;
         }
         System.out.println(sum % 10);
     }
